+++ syngo/__init__.py
@@ -1,6 +1,7 @@
 from datetime import datetime, timedelta
 
 from django.conf import settings
+from django.contrib.auth import get_user_model
 
 import httpx
 
@@ -11,20 +12,22 @@ def req(url, method="get", auth=True, **kwargs):
     return getattr(httpx, method)(f"{settings.SYNGO_MATRIX_URL}{url}", **kwargs)
 
 
-def django_to_matrix(user):
+def get_user_id(user):
     """Get the matrix id associated to a django user."""
-    return f"@{user.username}:{settings.SYNGO_MATRIX_DOMAIN}"
+    if isinstance(user, get_user_model()):
+        return f"@{user.username}:{settings.SYNGO_MATRIX_DOMAIN}"
+    return user
 
 
-def register(django_user):
+def register(user):
     """Register a Django user into a Matrix homeserver."""
     # https://matrix-org.github.io/synapse/latest/admin_api/user_admin_api.html
     # create-or-modify-account
-    user_id = django_to_matrix(django_user)
+    user_id = get_user_id(user)
     return req(
         f"/_synapse/admin/v2/users/{user_id}",
         method="put",
-        json={"displayname": str(django_user)},
+        json={"displayname": str(user)},
     )
 
 
@@ -49,22 +52,22 @@ def list_accounts(guests=False):
     return accounts
 
 
-def shadow_ban(django_user, unban=False):
+def shadow_ban(user, unban=False):
     """Shadow-(un)ban an user."""
     # https://matrix-org.github.io/synapse/latest/admin_api/user_admin_api.html
     # controlling-whether-a-user-is-shadow-banned
-    user_id = django_to_matrix(django_user)
+    user_id = get_user_id(user)
     return req(
         f"/_synapse/admin/v1/users/{user_id}/shadow_ban",
         method="delete" if unban else "post",
     )
 
 
-def deactivate(django_user):
+def deactivate(user):
     """Deactivate an account."""
     # https://matrix-org.github.io/synapse/latest/admin_api/user_admin_api.html
     # deactivate-account
-    user_id = django_to_matrix(django_user)
+    user_id = get_user_id(user)
     return req(
         f"/_synapse/admin/v1/deactivate/{user_id}", method="post", json={"erase": True}
     )